+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from "./certification";
 export * from "./errors";
 export * from "./provider";
 export * from "./security";